Refactoring ShowControlle, ShowService & Blade templates, few fix

+++ app/Data/Request/TaskDTO/CreateData.php
@@ -27,7 +27,7 @@ class CreateData implements RequestDataInterface
             priority: (int)data_get($data, 'priority'),
             title: data_get($data, 'title'),
             description: data_get($data, 'description'),
-            parent_id: (int)data_get($data, 'parent_id'),
+            parent_id: data_get($data, 'parent_id'),
         );
     }
